Add functionality for route protection
added two components, one for route Protection called RouteProtector, it sees if
user doesn't have userData saved to profile, and if so, it forwards the user to homepage.

the other component is called ForwardToFeed, if user is on homepage, login page, or register page,
and user has userData saved in their localStorage, it forwards the user to feed page.

also implemented logOut button in navbar for easy testing